refactor : 디렉토리명 정의
