+++ src/main/java/dbdr/global/util/api/ApiUtils.java
@@ -1,4 +1,4 @@
-package dbdr.global.util;
+package dbdr.global.util.api;
 
 import org.springframework.http.HttpStatus;
